compilation fixes for wxMSW
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@2797 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/generic/textdlgg.cpp
@@ -42,7 +42,7 @@ static wxSize wxSplitMessage2( const wxString &message, wxWindow *parent )
     int y = 10;
     int w = 50;
     wxString line( _T("") );
-    for (uint pos = 0; pos < message.Len(); pos++)
+    for (size_t pos = 0; pos < message.Len(); pos++)
     {
         if (message[pos] == _T('\n'))
         {
